fixed codeSmells and improved something
