--- src/utils/create-size.js
@@ -1,27 +1,0 @@
-import { config } from '../configs';
-
-const { languages } = config;
-
-export const createSizeNamelist = () => [
-  'heightInCm',
-  'widthInCm',
-  'depthInCm',
-  'volumeInLiters',
-  'weightInKg',
-  'additionalPrice'
-];
-
-export const createSize = (data) => ({
-  name: data.name,
-  simpleName: [
-    { lang: languages[0], value: data.simpleNameUa },
-    { lang: languages[1], value: data.simpleNameEn }
-  ],
-  heightInCm: data.heightInCm,
-  widthInCm: data.widthInCm,
-  depthInCm: data.depthInCm,
-  volumeInLiters: data.volumeInLiters,
-  weightInKg: data.weightInKg,
-  available: data.available,
-  additionalPrice: data.additionalPrice
-});
